#24 테스트 코드 작성 - BoardRepository (기본적인 CRUD) (#30)
* #24 테스트 코드 작성 - BoardRepository (기본적인 CRUD)

* #24 테스트 코드 작성 - BoardRepository (기본적인 CRUD) - Rollback(true) 삭제


BoardRepository가 잘 작동하는지 기본적인 CRUD만 테스트

- Create
- Retrieve (Read)
- Update
- Delete
- Board 엔티티 클래스 Long views -> int views로 수정했습니다.

+++ src/main/java/com/gabia/gyebalja/domain/Board.java
@@ -20,7 +20,7 @@ public class Board{
     @Column(columnDefinition = "TEXT")
     private String content;
 
-    private Long views;
+    private int views;
   
     @OneToMany(mappedBy = "board")
     private List<Likes> likes = new ArrayList<>();
@@ -40,7 +40,7 @@ public class Board{
     private List<Comment> comments = new ArrayList<>();
 
     @Builder
-    public Board(String title, String content, Long views, User user, Education education){
+    public Board(String title, String content, int views, User user, Education education){
         this.title = title;
         this.content = content;
         this.views = views;
